refactor(web): W10 slice 3 — four operations namespaces collapse into lib/operate (BI-AB17E1A8)
The pass §3.1-b table: "lib/ops/ vs lib/operate/ vs lib/operations/ vs
lib/operations-run/ — four namespaces, one concern". lib/operate (56
entries, the IT4IT S5.7 Operate domain home) absorbs the other three:

- lib/ops/{backlog-origin,operator-triage,platform-stack}(+tests)
- lib/operations/source-health.ts
- lib/operations-run/{reap,reap.test}.ts

Move-only via git mv; the three dissolved directories are gone. 8 import
specifiers rewritten (ops stack-currency page, components/ops x4,
lib/capacity/contracts, lib/twin/operations-snapshot,
lib/queue/functions/taskrun-watchdog). reap.ts's NEUTRAL-HOME RULE header
follows the move: lib/operate is still outside the collision hot-zones
the rule names (queue/tak/routing/govern), so the rule's intent holds in
the merged namespace.

Characterization: lib/operate + components/ops + ops route vitest
75 files / 778 tests green; per-file test-count diff vs the pre-move
baseline is EMPTY modulo the three renamed directory prefixes (51
operate-family files). Application-boundary guard unchanged: 8 contexts,
531 cross-context imports, 63 exceptions (none of the four namespaces
was a governed context).

Docs-Impact-Decision: internal namespace hygiene, no user-facing change.
Design-Grounding-Decision: grounded in
docs/architecture/2026-08-16-simplify-and-strengthen-architecture-pass.md
§3.1-b + §5 W10 (BI-AB17E1A8); substrate reviewed — lib/operate barrel
(no collisions with incoming names), reap.ts neutral-home rule
(EP-8DC217EB BET-10 plan §4), lifecycle relative-import depth preserved.
Process-Spine-Decision: executes a documented architecture-pass move
(§3.1-b operations namespace collapse); mechanical git mv + import
rewrites.
