+++ apps/web/components/ops/BacklogItemRow.tsx
@@ -12,7 +12,7 @@ import { startBacklogBuild } from "@/lib/actions/backlog-build";
 import { type BacklogItemWithRelations } from "@/lib/backlog";
 import { resolveBacklogBuildActionState } from "@/lib/backlog-build-action-state";
 import { AGENT_NAME_MAP } from "@/lib/agent-routing";
-import { backlogItemOrigin, BACKLOG_ORIGIN_LABEL } from "@/lib/ops/backlog-origin";
+import { backlogItemOrigin, BACKLOG_ORIGIN_LABEL } from "@/lib/operate/backlog-origin";
 import { backlogItemLifecycleLabel } from "@/lib/backlog-visibility";
 
 type Props = {
